requirements.txt
aiogram==3.0.0b8
asyncpg
python-dotenv